Updated trust store implementation (#353)
