+++ cmd/notation/cert_gen.go
@@ -11,6 +11,7 @@ import (
 	"github.com/notaryproject/notation-core-go/testhelper"
 	"github.com/notaryproject/notation-go/config"
 	"github.com/notaryproject/notation-go/dir"
+	"github.com/notaryproject/notation/internal/certificate"
 	"github.com/notaryproject/notation/internal/osutil"
 	"github.com/notaryproject/notation/internal/slices"
 	"github.com/notaryproject/notation/pkg/configutil"
@@ -75,7 +76,7 @@ func generateTestCert(opts *certGenerateTestOpts) error {
 	trust := opts.trust
 	// Add to the trust store
 	if trust {
-		if err := AddCertCore(certPath, "ca", name, true); err != nil {
+		if err := certificate.AddCertCore(certPath, "ca", name, true); err != nil {
 			return err
 		}
 	}
